added test and deployed script for nft

+++ Foundry-nft/src/BasicNft.sol
@@ -1,9 +1,10 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
+
 import {ERC721} from "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 
 contract BasicNft is ERC721 {
-    uint256 private s_tokenCounter; // counter for token id
+    uint256 private s_tokenCounter; // counter for nft  token id
     mapping(uint256 => string) private s_tokenIdtoUri;
 
     constructor() ERC721("mark52", "mrk") {
@@ -16,9 +17,7 @@ contract BasicNft is ERC721 {
         s_tokenCounter++;
     }
 
-    function tokenURI(
-        uint _tokenId
-    ) public view override returns (string memory) {
+    function tokenURI(uint256 _tokenId) public view override returns (string memory) {
         return s_tokenIdtoUri[_tokenId];
     }
 }
